trying to reduce warnings precision loss from type conversion
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@3000 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/air/air.h
@@ -460,7 +460,7 @@ AIR_EXPORT void airShuffle(unsigned int *buff, unsigned int N, int perm);
 AIR_EXPORT void airShuffle_r(airRandMTState *state,
                              unsigned int *buff, unsigned int N,
                              int perm);
-AIR_EXPORT char *airDoneStr(float start, float here, float end, char *str);
+AIR_EXPORT char *airDoneStr(double start, double here, double end, char *str);
 AIR_EXPORT double airTime();
 AIR_EXPORT double airCbrt(double);
 AIR_EXPORT double airSgnPow(double, double);
